Moves test utilities to nkms_eth, code reorg, cleanup, and documentation updates.
